requirements.txt
lightgbm==3.1.0
xgboost==1.2.1
catboost==0.24.3
tensorflow==2.3.1
tensorflow-addons==0.11.2
umap-learn==0.4.6
matplotlib_venn==0.11.6
japanize-matplotlib==1.1.3
plotly==4.13.0
omegaconf==2.0.5
hydra-core==1.0
mlflow==1.12.1
sweetviz==2.0.2
loguru==0.5.3
jupyterthemes==0.20.0
texthero==1.0.9
textdistance==4.2.0
nltk==3.5
gensim==3.8.3